+++ src/hooks/useHoldingMutation/useHoldingMutation.js
@@ -4,15 +4,15 @@ import {
 
 import { useOkapiKy } from '@folio/stripes/core';
 
-const useHoldingMutation = (options = {}) => {
-  const ky = useOkapiKy();
+const useHoldingMutation = (tenant, options = {}) => {
+  const ky = useOkapiKy({ tenant });
 
   const { mutateAsync } = useMutation({
     mutationFn: (holding) => {
       const kyMethod = holding.id ? 'put' : 'post';
       const kyPath = holding.id ? `inventory/holdings/${holding.id}` : 'holdings-storage/holdings';
 
-      return ky[kyMethod](kyPath, { json: holding });
+      return ky[kyMethod](kyPath, { json: holding }).json();
     },
     ...options,
   });