+++ src/hooks/useUpdateOwnership/useUpdateOwnership.js
@@ -2,14 +2,12 @@ import { useMutation } from 'react-query';
 
 import { useOkapiKy } from '@folio/stripes/core';
 
-const useUpdateOwnership = () => {
+const useUpdateOwnership = path => {
   const ky = useOkapiKy();
 
   const { mutateAsync } = useMutation({
     mutationFn: (body) => {
-      const kyPath = 'inventory/holdings/update-ownership';
-
-      return ky.post(kyPath, { json: body });
+      return ky.post(path, { json: body });
     },
   });